Program for copying files for ItStep Academy
This program in  a separate thread copies the file in blocks of 4096 bytes to the specified location. Displays the progress of the copy using the ProgressBar. User can specify paths to files using the keyboard or using the dialog the window that appears when you click the "File" buttons. Running copying occurs when you press the "Copy" button or
Enter keys in the text box, which indicates the path where to copy.